+++ packages/fast-crud/src/components/basic/fs-button.tsx
@@ -44,7 +44,8 @@ export default defineComponent({
     buttonProps: {
       type: Object,
       default: undefined
-    }
+    },
+    className: {}
   },
   setup(props, ctx) {
     const { ui } = useUi();
@@ -115,7 +116,7 @@ export default defineComponent({
             "is-thin": !props.text && !ctx.slots.default
           }
         },
-        { ...ctx.attrs, ...props.buttonProps }
+        { class: props.className, ...ctx.attrs, ...props.buttonProps }
       );
       if (iconProp) {
         // @ts-ignore