+++ packages/fast-crud/src/components/extends/fs-dict-tree.vue
@@ -1,8 +1,14 @@
 <template>
-  <component :is="ui.treeSelect.name" :[ui.treeSelect.options]="computedOptions" :placeholder="computedPlaceholder" />
+  <component
+    :is="ui.treeSelect.name"
+    ref="treeRef"
+    :[ui.treeSelect.options]="computedOptions"
+    :placeholder="computedPlaceholder"
+    v-bind="computedBinding"
+  />
 </template>
 <script lang="ts">
-import { computed, defineComponent } from "vue";
+import { computed, defineComponent, ref } from "vue";
 import { useDict } from "../../use/use-dict";
 import { useI18n } from "../../locale";
 import { useUi } from "../../use";
@@ -46,9 +52,20 @@ export default defineComponent({
     });
 
     let usedDict = useDict(props, ctx);
+
     const computedOptions = usedDict.createComputedOptions();
+
+    const computedBinding = computed(() => {
+      const dict = usedDict.getDict();
+      return ui.treeSelect.buildOptionKeysNameBinding({
+        label: dict.label,
+        value: dict.value,
+        children: dict.children
+      });
+    });
     return {
       ui,
+      computedBinding,
       computedPlaceholder,
       ...usedDict,
       computedOptions