+++ packages/fast-crud/src/components/extends/fs-table-select.vue
@@ -17,7 +17,7 @@
         <div :style="{ width: '100%', height: height || '60vh' }">
           <fs-crud ref="crudRef" v-bind="crudBinding">
             <template #header-top>
-              <div class="fs-table-select-current">
+              <div v-if="showCurrent !== false" class="fs-table-select-current">
                 当前选中：
                 <fs-values-format
                   ref="valuesFormatRef"
@@ -39,7 +39,7 @@
   </div>
 </template>
 <script lang="tsx" setup>
-import { Dict, useFs, useMerge, useUi } from "../../use";
+import { Dict, useCompute, useFs, useMerge, useUi } from "../../use";
 import { computed, nextTick, ref, Ref, watch } from "vue";
 import { CreateCrudOptions, DynamicallyCrudOptions } from "../../d";
 import _ from "lodash-es";
@@ -72,6 +72,7 @@ type FsTableSelectProps = {
 
   /**
    * 是否显示选择框
+   * 有时候你只是想要那个选择的Dialog，那么你可以隐藏select，然后自定义激活方式
    */
   showSelect?: boolean;
   /**
@@ -79,6 +80,10 @@ type FsTableSelectProps = {
    */
   dialog?: any;
 
+  /**
+   * 对话框中是否显示当前选中值
+   */
+  showCurrent?: boolean;
   /**
    * 当前选中值 fs-values-format组件 配置
    */
@@ -122,7 +127,8 @@ const props = withDefaults(defineProps<FsTableSelectProps>(), {
   select: undefined,
   crudOptionsOverride: undefined,
   valueType: "value",
-  showSelect: true
+  showSelect: true,
+  showCurrent: true
 });
 
 const slots = defineSlots<{
@@ -243,6 +249,7 @@ const override: DynamicallyCrudOptions = computed(() => {
     getPageData() {
       return crudBinding.value.data;
     },
+    useCompute: useCompute,
     multiple: props.multiple,
     selectedRowKeys,
     onSelectedKeysChanged: async (changed) => {