+++ packages/fast-crud/src/use/use-compute.ts
@@ -118,24 +118,24 @@ function doComputed(
   });
 }
 
-export class ComputeValue<T> implements ComputeRef<T> {
-  computeFn: ComputeFn<T>;
-  constructor(computeFn: ComputeFn<T>) {
+export class ComputeValue<T = any, R = any> implements ComputeRef<T, R> {
+  computeFn: ComputeFn<T, R>;
+  constructor(computeFn: ComputeFn<T, R>) {
     this.computeFn = computeFn;
   }
 }
 
-export function compute<T>(computeFn: (context: ComputeContext) => T): ComputeValue<T> {
-  return new ComputeValue<T>(computeFn);
+export function compute<T = any, R = any>(computeFn: ComputeFn<T, R>): ComputeRef<T, R> {
+  return new ComputeValue<T, R>(computeFn);
 }
 
 export type GetContextFn = () => any;
 
-export class AsyncComputeValue<T, P = any> implements AsyncComputeRef<T> {
-  watch?: (context: ScopeContext) => P;
-  asyncFn: (value: P, getContextFn: GetContextFn) => Promise<T>;
+export class AsyncComputeValue<RV, R = any, WV = any> implements AsyncComputeRef<RV, R> {
+  watch?: (context: ScopeContext<R>) => WV;
+  asyncFn: (value: WV, getContextFn: GetContextFn) => Promise<RV>;
   defaultValue?: any;
-  constructor(options: AsyncComputeRef<T>) {
+  constructor(options: AsyncComputeRef<RV, R>) {
     const { asyncFn, defaultValue } = options;
     this.watch = options.watch;
     this.asyncFn = asyncFn;
@@ -144,8 +144,8 @@ export class AsyncComputeValue<T, P = any> implements AsyncComputeRef<T> {
 
   buildAsyncRef(getContextFn: GetContextFn) {
     getContextFn = getContextFn || function () {};
-    const asyncRef: Ref<T> = ref(this.defaultValue);
-    const computedValue = computed<P>(() => {
+    const asyncRef: Ref<RV> = ref(this.defaultValue);
+    const computedValue = computed<WV>(() => {
       if (this.watch) {
         return this.watch(getContextFn());
       }
@@ -154,7 +154,7 @@ export class AsyncComputeValue<T, P = any> implements AsyncComputeRef<T> {
 
     watch(
       () => computedValue.value,
-      async (value: P) => {
+      async (value: WV) => {
         //执行异步方法
         asyncRef.value = await this.asyncFn(value, getContextFn());
       },
@@ -164,8 +164,10 @@ export class AsyncComputeValue<T, P = any> implements AsyncComputeRef<T> {
     return asyncRef;
   }
 }
-export function asyncCompute<T>(options: AsyncComputeRef<T>): AsyncComputeValue<T> {
-  return new AsyncComputeValue<T>(options);
+export function asyncCompute<RV = any, R = any, WV = any>(
+  options: AsyncComputeRef<RV, R, WV>
+): AsyncComputeRef<RV, R, WV> {
+  return new AsyncComputeValue<RV, R, WV>(options);
 }
 export function useCompute() {
   return {