+++ packages/fast-extends/src/editor/type/config.ts
@@ -7,5 +7,8 @@ export type FsEditorConfig = {
 export type FsEditorWang5Config = { editorConfig: IEditorConfig; toolbarConfig: IToolbarConfig };
 export const defaultConfig = {
   wangEditor: {},
-  wangEditor5: { editorConfig: {}, toolbarConfig: {} }
+  wangEditor5: {
+    editorConfig: {},
+    toolbarConfig: {}
+  }
 } as FsEditorConfig;