+++ docs/zh/guide/other/changelogs/packages/fast-admin/fs-admin-naive-ui/CHANGELOG.md
@@ -3,6 +3,27 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.20.2](https://github.com/jekip/naive-ui-admin/compare/v1.20.1...v1.20.2) (2024-03-21)
+
+### Bug Fixes
+
+* 修复搜索栏只有一个输入框时点回车 会刷新浏览器的bug ([677114d](https://github.com/jekip/naive-ui-admin/commit/677114d6847be8559b33c9b36ca19b8d160ecc3d))
+
+### Performance Improvements
+
+* fs-admin-antdv4 支持主题设置 ([bab0b30](https://github.com/jekip/naive-ui-admin/commit/bab0b30180025613f842f87db8264f7094560ed3))
+
+## [1.20.1](https://github.com/jekip/naive-ui-admin/compare/v1.20.0...v1.20.1) (2024-02-27)
+
+**Note:** Version bump only for package naive-ui-admin
+
+# [1.20.0](https://github.com/jekip/naive-ui-admin/compare/v1.19.3...v1.20.0) (2024-01-28)
+
+### Performance Improvements
+
+* 优化free模式，支持默认不激活 ([aeaf0a6](https://github.com/jekip/naive-ui-admin/commit/aeaf0a683ecc24dcb86036daea363f3019347299))
+* table-select支持隐藏选中label ([04791d5](https://github.com/jekip/naive-ui-admin/commit/04791d54eb5cae852f345207b90f75f85fad4eda))
+
 ## [1.19.3](https://github.com/jekip/naive-ui-admin/compare/v1.19.2...v1.19.3) (2023-12-15)
 
 **Note:** Version bump only for package naive-ui-admin